placed sigint in being of run action
