+++ src/Application.cxx
@@ -11,7 +11,6 @@
 #include <TRestGeant4PhysicsLists.h>
 #include <TRestRun.h>
 
-#include <csignal>
 #include <regex>
 #ifndef GEANT4_WITHOUT_G4RunManagerFactory
 #include <G4RunManagerFactory.hh>
@@ -272,14 +271,6 @@ Options ProcessCommandLineOptions(int argc, char* const argv[]) {
 
 }  // namespace CommandLineOptions
 
-int interruptSignalHandler(const int, void* ptr) {
-    // See https://stackoverflow.com/a/43400143/11776908
-    cout << "Stopping Run! Program was manually stopped by user (CTRL+C)!" << endl;
-    const auto manager = (SimulationManager*)(ptr);
-    manager->StopSimulation();
-    return 0;
-}
-
 constexpr const char* geometryName = "Geometry";
 
 void Application::Run(const CommandLineOptions::Options& options) {
@@ -447,8 +438,6 @@ void Application::Run(const CommandLineOptions::Options& options) {
         exit(1);
     }
     gGeoManager->Write(geometryName, TObject::kOverwrite);
-
-    signal(SIGINT, (void (*)(int))interruptSignalHandler);  // Add custom signal handler before simulation
 
     cout << "Number of events: " << nEvents << endl;
     if (nEvents > 0)  // batch mode
